add A3C FF/LSTM switch flag
