+++ a3c_display.py
@@ -4,8 +4,7 @@ import numpy as np
 import random
 
 from game_state import GameState
-#from game_ac_network import GameACNetwork
-from game_ac_lstm_network import GameACLSTMNetwork
+from game_ac_network import GameACFFNetwork, GameACLSTMNetwork
 from a3c_training_thread import A3CTrainingThread
 from rmsprop_applier import RMSPropApplier
 
@@ -16,6 +15,7 @@ from constants import RMSP_EPSILON
 from constants import RMSP_ALPHA
 from constants import GRAD_NORM_CLIP
 from constants import USE_GPU
+from constants import USE_LSTM
 
 def choose_action(pi_values):
   values = []
@@ -35,8 +35,10 @@ def choose_action(pi_values):
 # use CPU for display tool
 device = "/cpu:0"
 
-#global_network = GameACNetwork(ACTION_SIZE, device)
-global_network = GameACLSTMNetwork(ACTION_SIZE, -1, device)
+if USE_LSTM:
+  global_network = GameACLSTMNetwork(ACTION_SIZE, -1, device)
+else:
+  global_network = GameACFFNetwork(ACTION_SIZE, device)
 
 learning_rate_input = tf.placeholder("float")
