+++ a3c_training_thread.py
@@ -6,12 +6,12 @@ import random
 from accum_trainer import AccumTrainer
 from game_state import GameState
 from game_state import ACTION_SIZE
-#from game_ac_network import GameACNetwork
-from game_ac_lstm_network import GameACLSTMNetwork
+from game_ac_network import GameACFFNetwork, GameACLSTMNetwork
 
 from constants import GAMMA
 from constants import LOCAL_T_MAX
 from constants import ENTROPY_BETA
+from constants import USE_LSTM
 
 
 class A3CTrainingThread(object):
@@ -28,8 +28,11 @@ class A3CTrainingThread(object):
     self.learning_rate_input = learning_rate_input
     self.max_global_time_step = max_global_time_step
 
-    #self.local_network = GameACNetwork(ACTION_SIZE, device)
-    self.local_network = GameACLSTMNetwork(ACTION_SIZE, thread_index, device)
+    if USE_LSTM:
+      self.local_network = GameACLSTMNetwork(ACTION_SIZE, thread_index, device)
+    else:
+      self.local_network = GameACFFNetwork(ACTION_SIZE, device)
+
     self.local_network.prepare_loss(ENTROPY_BETA)
 
     # TODO: don't need accum trainer anymore with batch
@@ -98,7 +101,8 @@ class A3CTrainingThread(object):
 
     start_local_t = self.local_t
 
-    start_lstm_state = self.local_network.lstm_state_out
+    if USE_LSTM:
+      start_lstm_state = self.local_network.lstm_state_out
     
     # t_max times loop
     for i in range(LOCAL_T_MAX):
@@ -139,7 +143,8 @@ class A3CTrainingThread(object):
           
         self.episode_reward = 0
         self.game_state.reset()
-        self.local_network.reset_state()
+        if USE_LSTM:
+          self.local_network.reset_state()
         break
 
     R = 0.0
@@ -168,19 +173,27 @@ class A3CTrainingThread(object):
       batch_td.append(td)
       batch_R.append(R)
 
-    batch_si.reverse()
-    batch_a.reverse()
-    batch_td.reverse()
-    batch_R.reverse()
+    if USE_LSTM:
+      batch_si.reverse()
+      batch_a.reverse()
+      batch_td.reverse()
+      batch_R.reverse()
 
-    sess.run( self.accum_gradients,
-              feed_dict = {
-                self.local_network.s: batch_si,
-                self.local_network.a: batch_a,
-                self.local_network.td: batch_td,
-                self.local_network.r: batch_R,
-                self.local_network.initial_lstm_state: start_lstm_state,
-                self.local_network.step_size : [len(batch_a)] } )
+      sess.run( self.accum_gradients,
+                feed_dict = {
+                  self.local_network.s: batch_si,
+                  self.local_network.a: batch_a,
+                  self.local_network.td: batch_td,
+                  self.local_network.r: batch_R,
+                  self.local_network.initial_lstm_state: start_lstm_state,
+                  self.local_network.step_size : [len(batch_a)] } )
+    else:
+      sess.run( self.accum_gradients,
+                feed_dict = {
+                  self.local_network.s: batch_si,
+                  self.local_network.a: batch_a,
+                  self.local_network.td: batch_td,
+                  self.local_network.r: batch_R} )
       
     cur_learning_rate = self._anneal_learning_rate(global_t)
