+++ game_ac_network.py
@@ -1,49 +1,17 @@
 # -*- coding: utf-8 -*-
 import tensorflow as tf
 import numpy as np
+from custom_lstm import CustomBasicLSTMCell
 
-# Actor-Critic Network (Policy network and Value network)
+# Actor-Critic Network Base Class
+# (Policy network and Value network)
 
 class GameACNetwork(object):
   def __init__(self,
                action_size,
                device="/cpu:0"):
     self._device = device
-    
-    with tf.device(self._device):
-      self._action_size = action_size
-      
-      self.W_conv1 = self._conv_weight_variable([8, 8, 4, 16])  # stride=4
-      self.b_conv1 = self._conv_bias_variable([16], 8, 8, 4)
-
-      self.W_conv2 = self._conv_weight_variable([4, 4, 16, 32]) # stride=2
-      self.b_conv2 = self._conv_bias_variable([32], 4, 4, 16)
-
-      self.W_fc1 = self._fc_weight_variable([2592, 256])
-      self.b_fc1 = self._fc_bias_variable([256], 2592)
-
-      # weight for policy output layer
-      self.W_fc2 = self._fc_weight_variable([256, action_size])
-      self.b_fc2 = self._fc_bias_variable([action_size], 256)
-
-      # weight for value output layer
-      self.W_fc3 = self._fc_weight_variable([256, 1])
-      self.b_fc3 = self._fc_bias_variable([1], 256)
-
-      # state (input)
-      self.s = tf.placeholder("float", [None, 84, 84, 4])
-    
-      h_conv1 = tf.nn.relu(self._conv2d(self.s, self.W_conv1, 4) + self.b_conv1)
-      h_conv2 = tf.nn.relu(self._conv2d(h_conv1, self.W_conv2, 2) + self.b_conv2)
-
-      h_conv2_flat = tf.reshape(h_conv2, [-1, 2592])
-      h_fc1 = tf.nn.relu(tf.matmul(h_conv2_flat, self.W_fc1) + self.b_fc1)
-
-      # policy (output)
-      self.pi = tf.nn.softmax(tf.matmul(h_fc1, self.W_fc2) + self.b_fc2)
-      # value (output)
-      v_ = tf.matmul(h_fc1, self.W_fc3) + self.b_fc3
-      self.v = tf.reshape( v_, [-1] )
+    self._action_size = action_size
 
   def prepare_loss(self, entropy_beta):
     with tf.device(self._device):
@@ -70,23 +38,16 @@ class GameACNetwork(object):
       self.total_loss = policy_loss + value_loss
 
   def run_policy_and_value(self, sess, s_t):
-    pi_out, v_out = sess.run( [self.pi, self.v], feed_dict = {self.s : [s_t]} )
-    return (pi_out[0], v_out[0])
-
+    raise NotImplementedError()
+    
   def run_policy(self, sess, s_t):
-    pi_out = sess.run( self.pi, feed_dict = {self.s : [s_t]} )
-    return pi_out[0]
+    raise NotImplementedError()
 
   def run_value(self, sess, s_t):
-    v_out = sess.run( self.v, feed_dict = {self.s : [s_t]} )
-    return v_out[0]
+    raise NotImplementedError()    
 
   def get_vars(self):
-    return [self.W_conv1, self.b_conv1,
-            self.W_conv2, self.b_conv2,
-            self.W_fc1, self.b_fc1,
-            self.W_fc2, self.b_fc2,
-            self.W_fc3, self.b_fc3]
+    raise NotImplementedError()
 
   def sync_from(self, src_netowrk, name=None):
     src_vars = src_netowrk.get_vars()
@@ -130,3 +91,171 @@ class GameACNetwork(object):
 
   def _conv2d(self, x, W, stride):
     return tf.nn.conv2d(x, W, strides = [1, stride, stride, 1], padding = "VALID")
+
+# Actor-Critic FF Network
+
+class GameACFFNetwork(GameACNetwork):
+  def __init__(self,
+               action_size,
+               device="/cpu:0"):
+    GameACNetwork.__init__(self, action_size, device)
+    
+    with tf.device(self._device):
+      self.W_conv1 = self._conv_weight_variable([8, 8, 4, 16])  # stride=4
+      self.b_conv1 = self._conv_bias_variable([16], 8, 8, 4)
+
+      self.W_conv2 = self._conv_weight_variable([4, 4, 16, 32]) # stride=2
+      self.b_conv2 = self._conv_bias_variable([32], 4, 4, 16)
+
+      self.W_fc1 = self._fc_weight_variable([2592, 256])
+      self.b_fc1 = self._fc_bias_variable([256], 2592)
+
+      # weight for policy output layer
+      self.W_fc2 = self._fc_weight_variable([256, action_size])
+      self.b_fc2 = self._fc_bias_variable([action_size], 256)
+
+      # weight for value output layer
+      self.W_fc3 = self._fc_weight_variable([256, 1])
+      self.b_fc3 = self._fc_bias_variable([1], 256)
+
+      # state (input)
+      self.s = tf.placeholder("float", [None, 84, 84, 4])
+    
+      h_conv1 = tf.nn.relu(self._conv2d(self.s, self.W_conv1, 4) + self.b_conv1)
+      h_conv2 = tf.nn.relu(self._conv2d(h_conv1, self.W_conv2, 2) + self.b_conv2)
+
+      h_conv2_flat = tf.reshape(h_conv2, [-1, 2592])
+      h_fc1 = tf.nn.relu(tf.matmul(h_conv2_flat, self.W_fc1) + self.b_fc1)
+
+      # policy (output)
+      self.pi = tf.nn.softmax(tf.matmul(h_fc1, self.W_fc2) + self.b_fc2)
+      # value (output)
+      v_ = tf.matmul(h_fc1, self.W_fc3) + self.b_fc3
+      self.v = tf.reshape( v_, [-1] )
+
+  def run_policy_and_value(self, sess, s_t):
+    pi_out, v_out = sess.run( [self.pi, self.v], feed_dict = {self.s : [s_t]} )
+    return (pi_out[0], v_out[0])
+
+  def run_policy(self, sess, s_t):
+    pi_out = sess.run( self.pi, feed_dict = {self.s : [s_t]} )
+    return pi_out[0]
+
+  def run_value(self, sess, s_t):
+    v_out = sess.run( self.v, feed_dict = {self.s : [s_t]} )
+    return v_out[0]
+
+  def get_vars(self):
+    return [self.W_conv1, self.b_conv1,
+            self.W_conv2, self.b_conv2,
+            self.W_fc1, self.b_fc1,
+            self.W_fc2, self.b_fc2,
+            self.W_fc3, self.b_fc3]
+
+# Actor-Critic LSTM Network
+
+class GameACLSTMNetwork(GameACNetwork):
+  def __init__(self,
+               action_size,
+               thread_index, # -1 for global
+               device="/cpu:0" ):
+    GameACNetwork.__init__(self, action_size, device)    
+
+    with tf.device(self._device):
+      self.W_conv1 = self._conv_weight_variable([8, 8, 4, 16])  # stride=4
+      self.b_conv1 = self._conv_bias_variable([16], 8, 8, 4)
+
+      self.W_conv2 = self._conv_weight_variable([4, 4, 16, 32]) # stride=2
+      self.b_conv2 = self._conv_bias_variable([32], 4, 4, 16)
+
+      self.W_fc1 = self._fc_weight_variable([2592, 256])
+      self.b_fc1 = self._fc_bias_variable([256], 2592)
+
+      # lstm
+      self.lstm = CustomBasicLSTMCell(256)
+
+      # weight for policy output layer
+      self.W_fc2 = self._fc_weight_variable([256, action_size])
+      self.b_fc2 = self._fc_bias_variable([action_size], 256)
+
+      # weight for value output layer
+      self.W_fc3 = self._fc_weight_variable([256, 1])
+      self.b_fc3 = self._fc_bias_variable([1], 256)
+
+      # state (input)
+      self.s = tf.placeholder("float", [None, 84, 84, 4])
+    
+      h_conv1 = tf.nn.relu(self._conv2d(self.s, self.W_conv1, 4) + self.b_conv1)
+      h_conv2 = tf.nn.relu(self._conv2d(h_conv1, self.W_conv2, 2) + self.b_conv2)
+
+      h_conv2_flat = tf.reshape(h_conv2, [-1, 2592])
+      h_fc1 = tf.nn.relu(tf.matmul(h_conv2_flat, self.W_fc1) + self.b_fc1)
+      # h_fc1 shape=(5,256)
+
+      h_fc1_reshaped = tf.reshape(h_fc1, [1,-1,256])
+      # h_fc_reshaped = (1,5,256)
+
+      self.step_size = tf.placeholder(tf.float32, [1])
+
+      self.initial_lstm_state = tf.placeholder(tf.float32, [1, self.lstm.state_size])
+      
+      scope = "net_" + str(thread_index)
+
+      # time_major = False, so output shape is [batch_size, max_time, cell.output_size]
+      lstm_outputs, self.lstm_state = tf.nn.dynamic_rnn(self.lstm,
+                                                        h_fc1_reshaped,
+                                                        initial_state = self.initial_lstm_state,
+                                                        sequence_length = self.step_size,
+                                                        time_major = False,
+                                                        scope = scope)
+
+      # lstm_outputs: (1,5,256), (1,1,256)
+      
+      lstm_outputs = tf.reshape(lstm_outputs, [-1,256])
+
+      # policy (output)
+      self.pi = tf.nn.softmax(tf.matmul(lstm_outputs, self.W_fc2) + self.b_fc2)
+      
+      # value (output)
+      v_ = tf.matmul(lstm_outputs, self.W_fc3) + self.b_fc3
+      self.v = tf.reshape( v_, [-1] )
+
+      self.reset_state()
+      
+  def reset_state(self):
+    self.lstm_state_out = np.zeros([1, self.lstm.state_size])
+
+  def run_policy_and_value(self, sess, s_t):
+    pi_out, v_out, self.lstm_state_out = sess.run( [self.pi, self.v, self.lstm_state],
+                                                   feed_dict = {self.s : [s_t],
+                                                                self.initial_lstm_state : self.lstm_state_out,
+                                                                self.step_size : [1]} )
+    # pi_out: (1,3), v_out: (1)
+    return (pi_out[0], v_out[0])
+
+  def run_policy(self, sess, s_t):
+    pi_out, self.lstm_state_out = sess.run( [self.pi, self.lstm_state],
+                                            feed_dict = {self.s : [s_t],
+                                                         self.initial_lstm_state : self.lstm_state_out,
+                                                         self.step_size : [1]} )
+                                            
+    return pi_out[0]
+
+  def run_value(self, sess, s_t):
+    prev_lstm_state_out = self.lstm_state_out
+    v_out, _ = sess.run( [self.v, self.lstm_state],
+                         feed_dict = {self.s : [s_t],
+                                      self.initial_lstm_state : self.lstm_state_out,
+                                      self.step_size : [1]} )
+    
+    # roll back lstm state
+    self.lstm_state_out = prev_lstm_state_out
+    return v_out[0]
+
+  def get_vars(self):
+    return [self.W_conv1, self.b_conv1,
+            self.W_conv2, self.b_conv2,
+            self.W_fc1, self.b_fc1,
+            self.lstm.matrix, self.lstm.bias,
+            self.W_fc2, self.b_fc2,
+            self.W_fc3, self.b_fc3]
